# internal/apiserver/docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "IAM 系统 REST API 文档，包含认证(Authentication)、授权(Authorization)、身份管理(Identity)和身份提供商(IDP)模块",
        "title": "IAM API Documentation",
        "termsOfService": "https://iam.fangcunmount.cn/terms",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/FangcunMount/iam",
            "email": "support@fangcunmount.cn"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0.0"
    },
    "host": "iam.fangcunmount.cn",
    "basePath": "/api",
    "paths": {
...
        "/v2/authn/logout": {
            "post": {
                "description": "撤销访问令牌和刷新令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "用户登出",
                "parameters": [
                    {
                        "description": "登出请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "登出成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/v2/authn/refresh_token": {
            "post": {
                "description": "使用刷新令牌获取新的访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "刷新访问令牌",
                "parameters": [
                    {
                        "description": "刷新令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.RefreshTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "刷新成功，返回新的访问令牌",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenPair"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "刷新令牌无效或已过期",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/v2/authn/verify": {
            "post": {
                "description": "验证访问令牌的有效性并返回声明信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "认证"
                ],
                "summary": "验证访问令牌",
                "parameters": [
                    {
                        "description": "验证令牌请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.VerifyTokenRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "验证成功",
                        "schema": {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenVerifyResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "令牌无效",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.LogoutRequest": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "可选，撤销访问令牌",
                    "type": "string"
                },
                "refresh_token": {
                    "description": "可选，撤销刷新令牌",
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.RefreshTokenRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_request.VerifyTokenRequest": {
            "type": "object",
            "required": [
                "access_token"
            ],
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expected_audience": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expected_issuer": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenClaims": {
            "type": "object",
            "properties": {
                "amr": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "attributes": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "audience": {
                    "description": "受众",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "authenticated_at": {
                    "description": "原始认证时间",
                    "type": "string"
                },
                "expires_at": {
                    "description": "过期时间",
                    "type": "string"
                },
                "issued_at": {
                    "description": "签发时间",
                    "type": "string"
                },
                "issuer": {
                    "description": "签发者",
                    "type": "string"
                },
                "jti": {
                    "description": "JWT ID（可选）",
                    "type": "string"
                },
                "login_identity_id": {
                    "description": "登录身份 ID",
                    "type": "string"
                },
                "not_before": {
                    "description": "生效时间",
                    "type": "string"
                },
                "org_id": {
                    "description": "业务组织 ID",
                    "type": "string"
                },
                "session_id": {
                    "description": "会话 ID",
                    "type": "string"
                },
                "subject": {
                    "description": "主体",
                    "type": "string"
                },
                "tenant_domain": {
                    "description": "IAM 授权域",
                    "type": "string"
                },
                "tenant_id": {
                    "description": "历史租户数值 ID（兼容字段）",
                    "type": "integer"
                },
                "token_type": {
                    "description": "access/service",
                    "type": "string"
                },
                "user_id": {
                    "description": "用户 ID",
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "expires_in": {
                    "description": "过期时间（秒）",
                    "type": "integer"
                },
                "refresh_token": {
                    "description": "刷新令牌（可选）",
                    "type": "string"
                },
                "token_type": {
                    "description": "令牌类型（Bearer）",
                    "type": "string"
                }
            }
        },
        "github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenVerifyResponse": {
            "type": "object",
            "properties": {
                "claims": {
                    "description": "令牌声明（如果有效）",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_FangcunMount_iam_v2_internal_apiserver_transport_rest_authn_response.TokenClaims"
                        }
                    ]
                },
                "valid": {
                    "description": "令牌是否有效",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT 认证令牌，格式: Bearer {access_token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}